fix for missing fonts and “font-display: optional”
this is really interesting: optional: The new kid on the block, and possibly a good solution for all our font loading problems. With font-display: optional, there’s a window of only about 100ms for the custom font to load (the block period). If the font doesn’t load during that block period, the fallback font is used and the custom font is not loaded at all, but it’s still downloaded and cached behind the scenes. This means that, on subsequent page loads, the custom font should be available in the cache and will load instantly.